Info area sizes: allow fractional mm (e.g. 27,5) with comma or dot input
Operators couldn't save a size like 27,5 mm: the columns were Int, the zod
schemas required .int(), and the number inputs rejected a comma decimal.

- Schema: InfoAreaSize widthMm/heightMm Int → Float; migration
  20260615170000_info_area_sizes_decimal widens the columns to double
  precision (guarded DO block — lossless on existing integer rows, safe to
  re-run). Railway applies it on deploy start before the new code serves.
- API zod (info-area-sizes create/[id] PATCH + per-output size PATCH): drop
  .int(), keep positive ≤1000.
- UI: width/height inputs are now decimal-friendly text inputs; a comma or
  dot is normalised to a number (parseMm), and sizes render with a comma
  (fmtMm) in the settings list, the style-card readout and the size dropdown.

Verified: tsc clean; parse/format logic unit-checked; dev server reads the
existing prod table via the Float client with no errors (settings page 200).

+++ src/app/api/admin/info-area-sizes/[id]/route.ts
@@ -7,8 +7,9 @@ export const runtime = "nodejs";
 
 const PATCH_SCHEMA = z.object({
   name: z.string().min(1).max(120).optional(),
-  widthMm: z.number().int().positive().max(1000).optional(),
-  heightMm: z.number().int().positive().max(1000).optional(),
+  // Fractional mm allowed (e.g. 27.5) — see route.ts.
+  widthMm: z.number().positive().max(1000).optional(),
+  heightMm: z.number().positive().max(1000).optional(),
   active: z.boolean().optional(),
 });
